[Anagrammer] Stop when word lengths are consumed
Currently we find results that consume the entire query, even if the sum
of the word lengths is strictly smaller, e.g. {query: "AT.", wordLengths:
[2]} will return "AT A", "AT I", etc. instead of just "AT".

I've never found the current behavior useful. However, the downside is
that fewer options for the first few words (with word lengths specified)
are returned, if there are many possibilities for the later words.